Switch window/display lib to sdl2

+++ src/resources.rs
@@ -16,7 +16,7 @@ pub struct Height(pub usize);
 pub struct FrameCount(pub u32);
 
 #[derive(Debug, Default)]
-pub struct BufferOutput(pub Vec<u32>);
+pub struct BufferOutput(pub Vec<u8>);
 
 #[derive(Debug, Default)]
 pub struct ImageFilePrefix(pub String);
